ISSUE #526 rename file for consistency
