+++ test/src/unit/repo/lib/CMakeLists.txt
@@ -8,8 +8,8 @@ set(TEST_SOURCES
 	${TEST_SOURCES}
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_config.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_matrix.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_time_zone_converter.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_uuid.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/ut_repo_vector2d.cpp
-	${CMAKE_CURRENT_SOURCE_DIR}/ut_time_zone_converter.cpp
 	CACHE STRING "TEST_SOURCES" FORCE)
